working now

+++ Project/DoubleExtension.swift
@@ -8,10 +8,10 @@
 
 import Foundation
 
-extension Float {
+extension Double {
     /// Rounds the double to decimal places value
-    func roundTo(places:Int) -> Float {
-        let divisor = pow(10.0, Float(places))
+    func roundTo(places:Int) -> Double {
+        let divisor = pow(10.0, Double(places))
         return (self * divisor).rounded() / divisor
     }
 }
